Refactor integration testing to allow for better login flow testing
